requirements.txt
pyside6
pillow
pydicom
numpy
tifffile
matplotlib
scikit-learn
scikit-image
opencv-python
pymedphys
imagecodecs
pylibjpeg 
pylibjpeg-libjpeg 
pylibjpeg-openjpeg
interpolation
mkdocs
mkdocs-material
mkdocstrings[python]
pyqtgraph